# openapi.json
{"openapi":"3.1.0","info":{"title":"Velra","version":"1.0.0"},"paths":{"/auth/signup":{"post":{"tags":["auth"],"summary":"Signup","operationId":"signup_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/analyze-chat":{"post":{"tags":["chat-analysis"],"summary":"Analyze Chat","operationId":"analyze_chat_analyze_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/history":{"get":{"tags":["chat-analysis"],"summary":"Get History","operationId":"get_history_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AnalyzeChatResponse"},"type":"array","title":"Response Get History History Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/feedback":{"post":{"tags":["feedback"],"summary":"Submit Feedback","operationId":"submit_feedback_feedback_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/profile/analyze":{"post":{"tags":["profiles"],"summary":"Analyze Profile","operationId":"analyze_profile_profile_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileAnalyzeRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/health":{"get":{"tags":["system"],"summary":"Healthcheck","operationId":"healthcheck_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Healthcheck Health Get"}}}}}}}},"components":{"schemas":{"AnalyzeChatRequest":{"properties":{"messages":{"items":{"type":"string"},"type":"array","maxItems":200,"minItems":1,"title":"Messages"},"context":{"type":"string","maxLength":64,"minLength":2,"title":"Context"},"profile_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Id"}},"type":"object","required":["messages","context"],"title":"AnalyzeChatRequest"},"AnalyzeChatResponse":{"properties":{"seriousness_score":{"type":"integer","title":"Seriousness Score"},"interest_level":{"type":"string","enum":["low","medium","high","declining"],"title":"Interest Level"},"behavioral_pattern":{"items":{"type":"string"},"type":"array","title":"Behavioral Pattern"},"emotional_investment":{"type":"string","title":"Emotional Investment","default":"moderate"},"risk_level":{"type":"string","title":"Risk Level","default":"medium"},"date_strategy":{"$ref":"#/components/schemas/DateStrategy"},"impression_strategy":{"items":{"type":"string"},"type":"array","title":"Impression Strategy","default":[]},"suggestions":{"items":{"type":"string"},"type":"array","title":"Suggestions","default":[]},"replies":{"items":{"type":"string"},"type":"array","title":"Replies","default":[]},"effort_level":{"type":"string","title":"Effort Level","default":"balanced"}},"type":"object","required":["seriousness_score","interest_level","behavioral_pattern","date_strategy"],"title":"AnalyzeChatResponse"},"DateStrategy":{"properties":{"budget":{"type":"string","title":"Budget","default":"$$"},"type":{"type":"string","title":"Type","default":"coffee"},"justification":{"type":"string","title":"Justification","default":"Standard fallback recommendation"}},"type":"object","title":"DateStrategy"},"FeedbackRequest":{"properties":{"analysis_id":{"type":"string","title":"Analysis Id"},"is_helpful":{"type":"boolean","title":"Is Helpful"}},"type":"object","required":["analysis_id","is_helpful"],"title":"FeedbackRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProfileAnalyzeRequest":{"properties":{"text":{"type":"string","maxLength":5000,"minLength":10,"title":"Text"},"source":{"type":"string","enum":["instagram","linkedin","manual"],"title":"Source"}},"type":"object","required":["text","source"],"title":"ProfileAnalyzeRequest"},"ProfileDetailResponse":{"properties":{"id":{"type":"string","title":"Id"},"source":{"type":"string","title":"Source"},"extracted_traits":{"$ref":"#/components/schemas/ProfileTraitsResponse"}},"type":"object","required":["id","source","extracted_traits"],"title":"ProfileDetailResponse"},"ProfileTraitsResponse":{"properties":{"personality_type":{"type":"string","title":"Personality Type"},"interests":{"items":{"type":"string"},"type":"array","title":"Interests"},"lifestyle":{"type":"string","title":"Lifestyle"},"communication_style":{"type":"string","title":"Communication Style"},"values":{"items":{"type":"string"},"type":"array","title":"Values"},"likely_preferences":{"items":{"type":"string"},"type":"array","title":"Likely Preferences"}},"type":"object","required":["personality_type","interests","lifestyle","communication_style","values","likely_preferences"],"title":"ProfileTraitsResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}